#98: Move test packages
