+++ editor/core/build.gradle.kts
@@ -24,6 +24,6 @@ dependencies {
 tasks.register<JavaExec>("e2e") {
 	group = "verification"
 
-	mainClass.set("e2e.E2EKt")
+	mainClass.set("dev.kkorolyov.pancake.editor.core.test.e2e.E2EKt")
 	classpath = sourceSets.test.get().runtimeClasspath
 }
