+++ editor/src/test/kotlin/dev/kkorolyov/pancake/editor/test/e2e/E2E.kt
@@ -1,0 +1,20 @@
+package dev.kkorolyov.pancake.editor.test.e2e
+
+import dev.kkorolyov.pancake.editor.test.drawEnd
+import dev.kkorolyov.pancake.editor.test.drawStart
+import dev.kkorolyov.pancake.editor.test.editor
+import dev.kkorolyov.pancake.editor.test.start
+import dev.kkorolyov.pancake.platform.GameEngine
+import dev.kkorolyov.pancake.platform.Pipeline
+
+fun main() {
+	start(GameEngine().apply {
+		setPipelines(
+			Pipeline(
+				drawStart(),
+				editor(this),
+				drawEnd()
+			)
+		)
+	})
+}
