--- editor/src/test/kotlin/e2e/E2E.kt
@@ -1,16 +1,0 @@
-package e2e
-
-import dev.kkorolyov.pancake.platform.GameEngine
-import dev.kkorolyov.pancake.platform.Pipeline
-
-fun main() {
-	start(GameEngine().apply {
-		setPipelines(
-			Pipeline(
-				drawStart(),
-				editor(this),
-				drawEnd()
-			)
-		)
-	})
-}
